fix PR files
