+++ compose.yml
@@ -33,3 +33,4 @@ volumes:
 
 networks:
   backend: { }
+
