+++ scrapper/src/test/java/edu/java/scrapper/integrationTests/IntegrationTest.java
@@ -1,21 +1,5 @@
-package edu.java.scrapper;
+package edu.java.scrapper.integrationTests;
 
-import liquibase.Contexts;
-import liquibase.LabelExpression;
-import liquibase.Liquibase;
-import liquibase.database.Database;
-import liquibase.database.DatabaseFactory;
-import liquibase.database.jvm.JdbcConnection;
-import liquibase.exception.DatabaseException;
-import liquibase.exception.LiquibaseException;
-import liquibase.resource.DirectoryResourceAccessor;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.JdbcDatabaseContainer;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Testcontainers;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.nio.file.Path;
@@ -27,6 +11,21 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
+import liquibase.Contexts;
+import liquibase.LabelExpression;
+import liquibase.Liquibase;
+import liquibase.database.Database;
+import liquibase.database.DatabaseFactory;
+import liquibase.database.jvm.JdbcConnection;
+import liquibase.exception.LiquibaseException;
+import liquibase.resource.DirectoryResourceAccessor;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.JdbcDatabaseContainer;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Testcontainers;
 
 @Testcontainers
 public class IntegrationTest {
@@ -60,7 +59,7 @@ public class IntegrationTest {
             new JdbcConnection(connection)
         );
 
-        Path path = new File(".").toPath().toAbsolutePath().getParent().getParent().resolve("migrations");
+        Path path = new File(".").toPath().toAbsolutePath().resolve("src").resolve("main").resolve("resources").resolve("migrations");
         Liquibase liquibase = new Liquibase(
             "master.xml",
             new DirectoryResourceAccessor(path),
@@ -74,29 +73,5 @@ public class IntegrationTest {
         registry.add("spring.datasource.url", POSTGRES::getJdbcUrl);
         registry.add("spring.datasource.username", POSTGRES::getUsername);
         registry.add("spring.datasource.password", POSTGRES::getPassword);
-    }
-
-    @Test
-    public void checkDatabase() throws SQLException {
-        Properties properties = new Properties();
-        properties.put("user", POSTGRES.getUsername());
-        properties.put("password", POSTGRES.getPassword());
-        Connection connection = DriverManager.getConnection(
-            POSTGRES.getJdbcUrl(),
-            properties
-        );
-
-        DatabaseMetaData databaseMetaData = connection.getMetaData();
-        ResultSet resultSet = databaseMetaData
-            .getTables(null, null, null, new String[]{"TABLE"});
-        List<String> tables = new ArrayList<>();
-        while (resultSet.next()) {
-            tables.add(resultSet.getString("TABLE_NAME"));
-        }
-
-        Assertions.assertEquals(
-            tables,
-            List.of("chat", "connections", "databasechangelog", "databasechangeloglock", "link")
-        );
     }
 }
